feat 0.0.13.3: added validation

+++ src/main/java/com/senla/courses/service/messages/MessageServiceImpl.java
@@ -6,6 +6,7 @@ import com.senla.courses.dao.TeacherDAO;
 import com.senla.courses.dao.UserDAO;
 import com.senla.courses.dto.MessageDTO;
 import com.senla.courses.dto.MessageFullDTO;
+import com.senla.courses.exception.ValidationException;
 import com.senla.courses.mapper.MessageMapper;
 import com.senla.courses.exception.EmptyListException;
 import com.senla.courses.exception.NotFoundException;
@@ -24,7 +25,7 @@ import java.util.Optional;
 
 @Service
 @AllArgsConstructor
-public class MessageServiceImpl implements MessageService{
+public class MessageServiceImpl implements MessageService {
 
     private final MessageMapper messageMapper;
     private final StudentDAO studentDAO;
@@ -65,9 +66,15 @@ public class MessageServiceImpl implements MessageService{
     }
 
     @Override
-    public MessageFullDTO updateMessage(MessageDTO messageDTO, Long id) {
+    public MessageFullDTO updateMessage(User userSec, MessageDTO messageDTO, Long id) {
         Message messageIn = messageDAO.find(id)
                 .orElseThrow(() -> new NotFoundException("Не смогли найти сообщение для обновления"));
+        Long from = userDAO.findByLogin(userSec.getUsername()).orElseThrow(()
+                -> new ValidationException("Не пройдена аутентификация") {
+        }).getId();
+        if (!messageIn.getFrom().getId().equals(from)) {
+            throw new ValidationException("Не пройдена аутентификация");
+        }
         messageIn.setBody(messageDTO.getBody());
         Message messageOut = messageDAO.update(messageIn);
         return messageMapper.fromMessage(messageOut);
@@ -81,7 +88,13 @@ public class MessageServiceImpl implements MessageService{
     }
 
     @Override
-    public List<MessageFullDTO> getMessagesBetween(List<Long> betweenIds, int from, int size) {
+    public List<MessageFullDTO> getMessagesBetween(User userSec, List<Long> betweenIds, int from, int size) {
+        Long fromValid = userDAO.findByLogin(userSec.getUsername()).orElseThrow(()
+                -> new ValidationException("Не пройдена аутентификация") {
+        }).getId();
+        if (!betweenIds.contains(fromValid)) {
+            throw new ValidationException("Можно просматривать только сообщения, в которых вы один из адресатов");
+        }
         List<Message> messagesList = messageDAO.findMessagesBetween(betweenIds.get(0), betweenIds.get(1), from, size);
         if (messagesList.isEmpty()) {
             throw new EmptyListException("Список сообщений между указанными id пусть");
@@ -90,8 +103,11 @@ public class MessageServiceImpl implements MessageService{
     }
 
     @Override
-    public List<MessageFullDTO> findMessagesByText(String text, int from, int size) {
-        List<Message> messagesList = messageDAO.findAllByText(text, from, size);
+    public List<MessageFullDTO> findMessagesByText(User userSec, String text, int from, int size) {
+        Long fromOrTo = userDAO.findByLogin(userSec.getUsername())
+                .orElseThrow(() -> new ValidationException("Не пройдена аутентификация") {
+                }).getId();
+        List<Message> messagesList = messageDAO.findAllByText(fromOrTo, text, from, size);
         if (messagesList.isEmpty()) {
             throw new EmptyListException("Список сообщений между указанными id пусть");
         }
@@ -99,7 +115,16 @@ public class MessageServiceImpl implements MessageService{
     }
 
     @Override
-    public void deleteMessage(Long id) {
-        messageDAO.deleteById(id);
+    public void deleteMessage(User userSec, Long id) {
+        Message message = messageDAO.find(id)
+                .orElseThrow(() -> new NotFoundException("Не смогли найти сообщение"));
+        Long from = userDAO.findByLogin(userSec.getUsername())
+                .orElseThrow(()
+                        -> new ValidationException("Не пройдена аутентификация") {
+                }).getId();
+        if (!message.getFrom().getId().equals(from)) {
+            throw new ValidationException("Можно удалять только сообщения, в которых вы -- отправитель");
+        }
+            messageDAO.deleteById(id);
     }
 }
