+++ src/main/java/com/senla/courses/service/teachers/TeacherServiceImpl.java
@@ -4,8 +4,10 @@ import com.senla.courses.dao.TeacherDAO;
 import com.senla.courses.dao.UserDAO;
 import com.senla.courses.dto.UserDTO;
 import com.senla.courses.exception.NotFoundException;
+import com.senla.courses.exception.ValidationException;
 import com.senla.courses.mapper.UserMapper;
 import com.senla.courses.exception.EmptyListException;
+import com.senla.courses.model.Role;
 import com.senla.courses.model.Teacher;
 import com.senla.courses.model.User;
 import lombok.AllArgsConstructor;
@@ -39,9 +41,12 @@ public class TeacherServiceImpl implements TeacherService {
     }
 
     @Override
-    public UserDTO updateTeacher(UserDTO userDTO, Long id) {
+    public UserDTO updateTeacher(org.springframework.security.core.userdetails.User userSec, UserDTO userDTO, Long id) {
         Teacher teacherUpd = teacherDAO.find(id)
                 .orElseThrow(() -> new NotFoundException("Не смогли найти учителя по id " + id));
+        if (teacherValidate(teacherUpd, userSec)) {
+            throw new ValidationException("Обновить можно только инфомрацию о себе");
+        }
         User user = userMapper.updateUser(teacherUpd.getUser(), userDTO);
         teacherUpd.setUser(user);
         teacherDAO.update(teacherUpd);
@@ -69,8 +74,18 @@ public class TeacherServiceImpl implements TeacherService {
 
 
     @Override
-    public void deleteTeacher(Long id) {
+    public void deleteTeacher(org.springframework.security.core.userdetails.User userSec, Long id) {
+        Teacher teacher = teacherDAO.find(id)
+                .orElseThrow(() -> new NotFoundException("Не смогли найти учителя по id " + id));
+        if (teacherValidate(teacher, userSec)) {
+            throw new ValidationException("Удалить можно только информацию о себе");
+        }
         teacherDAO.deleteById(id);
+    }
+
+    private boolean teacherValidate(Teacher teacher, org.springframework.security.core.userdetails.User userSec) {
+        return !teacher.getUser().getLogin().equals(userSec.getUsername())
+                && !userSec.getAuthorities().equals(Role.ADMIN.getAuthorities());
     }
 
 }
